+++ packages/javascript/lib/misc/index.ts
@@ -1,4 +1,5 @@
 export { debounce } from './debounce.js';
 export { funcTimes } from './funcTimes.js';
+export { retry } from './retry.js';
 export { sleep } from './sleep.js';
 export { throttle } from './throttle.js';